adds more examples
