+++ sdk/appconfiguration/examples/on_off_feature_manager.rs
@@ -1,0 +1,35 @@
+use appconfiguration::prelude::*;
+use azure_identity::DefaultAzureCredentialBuilder;
+use std::{env, sync::Arc};
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    env::set_var("AZCONFIG_NAME", "azure-rust-sdk");
+    // env::set_var("FEATURE_FETCH_ALL_OFF", "off");
+
+    let creds = Arc::new(
+        DefaultAzureCredentialBuilder::new()
+            .exclude_managed_identity_credential()
+            .build(),
+    );
+
+    let features = FeatureManager::new(creds, None);
+    println!("Features {features:?}");
+
+    println!("***on_off***");
+    let user = String::from("test");
+    let feature_name = String::from("on_off");
+
+    println!(
+        "Feature - {:?} is {:?} for User - {:?}",
+        feature_name,
+        if features.is_enabled(feature_name.clone()) {
+            "enabled"
+        } else {
+            "not enabled"
+        },
+        user,
+    );
+
+    Ok(())
+}
