+++ sdk/appconfiguration/src/prelude.rs
@@ -1,1 +1,3 @@
-pub use crate::feature_manager::*;
+pub use crate::feature_manager::{
+    app_context::AppContext, app_context::ContextHolder, FeatureManager,
+};
